Adding FLIP_C3 board (#112)
* Adding FLIP_C3 board

The FLIP_C3 is a simple ESP32-C3 board that has an integrated power supply that has a 60V input.  This board can be connected directly to a 48V battery and then be used with the M5Stack Isolated RS485 Unit (U094) to get a RS485 connection back to the master unit.  I successfully used 4 of these units to communicate with the JK-BMS B-Series BMS's.  The FLIP_C3 has an onboard 5V regulator that can power the RS485 module, 2 UART exposed on connectors that can be used for TTL to the BMS and TTL to the RS485 module. In the US it is available on Amazon, product page: https://wiki.vdbx.io/product/flip_c3

* Added missing GPIOs to FLIP_C3 board