Rename the switch "Force Bulk (top bal)" to "Force Bulk+Balancing"
For me it was not clear that the Force Bulk switch waited for balancing to start.
It was in the name (top bal), but I set Bulk equal with the balancing process, which it is not.

I think renaming it makes it clearer that this switches does something more than just setting the current charge instruction to Bulk as seen the the Force Bulk Logic:

        // +-----------------------------------------------+
        // | Force Bulk Logic                              |
        // +-----------------------------------------------+
        if (id(${yambms_id}_switch_force_bulk).state)
        {
          // Stop Force Bulk when the BMS starts to equalize the cells
          if (id(${yambms_id}_equalizing).state)
          {
            id(${yambms_id}_charge_status) = "Balancing";
            id(${yambms_id}_switch_force_bulk).turn_off();
          }
          // Force Bulk
          else id(${yambms_id}_charge_status) = "Force Bulk";
        }